requirements.txt
discord.py
google-cloud-firestore
google-cloud-vision
google-api-python-client
deepl
